Remove old with-amp example and rename experimental-amp to with-amp
