--- examples/experimental-amp/components/Byline.js
@@ -1,11 +1,0 @@
-export default ({ author }) => (
-  <>
-    <div className='byline'>By {author}</div>
-    <style jsx>{`
-      .byline {
-        color: green;
-        font-weight: bolder;
-      }
-    `}</style>
-  </>
-)
