+++ examples/with-amp/components/Byline.js
@@ -1,1 +1,11 @@
-export default ({ author }) => <div className='byline'>By {author}</div>
+export default ({ author }) => (
+  <>
+    <div className='byline'>By {author}</div>
+    <style jsx>{`
+      .byline {
+        color: green;
+        font-weight: bolder;
+      }
+    `}</style>
+  </>
+)
