+++ examples/with-amp/components/Layout.js
