+++ examples/with-amp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "with-amp",
+  "name": "experimental-amp",
   "version": "1.0.0",
   "scripts": {
     "dev": "next",
