+++ examples/with-amp/pages/normal.js
